wip testOutdatedSchemaFromClientThrowsExceptionOnServer
